Remove app assets in favor of global styles and scripts
