+++ frontend/module/main.ts
@@ -24,10 +24,26 @@
 //
 // See doc/COPYRIGHT.rdoc for more details.
 
+/**
+ * This file will be symlinked into the angular compiler project
+ * at the OpenProject core frontend folder.
+ *
+ * From here, you can work with anything angular related just in the core.
+ * When developing, we recommend you run the script `./bin/setup_dev` in the OpenProject
+ * core first, and then develop the frontend in the frontend/src/app/modules/plugins/linked/ folder.
+ *
+ * This will allow your IDE to pick up the angular project and provide import assistance and so on.
+ */
+
 import {APP_INITIALIZER, Injector, NgModule} from '@angular/core';
 import {CommonModule} from "@angular/common";
 import {KittenComponent} from "core-app/modules/plugins/linked/openproject-proto_plugin/kitten-component/kitten.component";
 import {HookService} from "core-app/modules/plugins/hook-service";
+
+import './global_scripts'
+import {KITTEN_ROUTES} from 'core-app/modules/plugins/linked/openproject-proto_plugin/kitten.routes';
+import {UIRouterModule} from '@uirouter/angular';
+import {KittenPageComponent} from 'core-app/modules/plugins/linked/openproject-proto_plugin/kitten-page/kitten-page.component';
 
 export function initializeProtoPlugin(injector:Injector) {
   return () => {
@@ -47,6 +63,7 @@ export function initializeProtoPlugin(injector:Injector) {
 @NgModule({
   imports: [
     CommonModule,
+    UIRouterModule.forChild({ states: KITTEN_ROUTES })
   ],
   providers: [
     // This initializer gets called when the Angular frontend is being loaded by the core
@@ -55,13 +72,9 @@ export function initializeProtoPlugin(injector:Injector) {
   ],
   declarations: [
     // Declare the component for angular to use
-    KittenComponent
+    KittenComponent,
+    KittenPageComponent,
   ],
-  entryComponents: [
-    // Special case: Declare the component also as a bootstrap component
-    // as it is being rendered from Rails.
-    KittenComponent
-  ]
 })
 export class PluginModule {
 }
